[FIX] fix type
- change type from "unsigned long long" to "long long"
- it is probably a bad idea to take the difference of unsigned integers
  as done here (ints[2] - extrapol) and store the result in an unsigned
  integer variable. The result may be negative and lead to unintended
  side effects. It is probably safer to use signed integers for these
  types of calculations all the way.

+++ src/main/cpp/MSNumpress.cpp
@@ -234,7 +234,7 @@ size_t encodeLinear(
 		unsigned char *result,
 		double fixedPoint
 ) {
-	unsigned long long ints[3];
+	long long ints[3];
 	size_t i, ri;
 	unsigned char halfBytes[10];
 	size_t halfByteCount;
@@ -248,14 +248,14 @@ size_t encodeLinear(
 
 	if (dataSize == 0) return 8;
 
-	ints[1] = data[0] * fixedPoint + 0.5;
+	ints[1] = static_cast<long long>(data[0] * fixedPoint + 0.5);
 	for (i=0; i<4; i++) {
 		result[8+i] = (ints[1] >> (i*8)) & 0xff;
 	}
 
 	if (dataSize == 1) return 12;
 
-	ints[2] = data[1] * fixedPoint + 0.5;
+	ints[2] = static_cast<long long>(data[1] * fixedPoint + 0.5);
 	for (i=0; i<4; i++) {
 		result[12+i] = (ints[2] >> (i*8)) & 0xff;
 	}
@@ -266,7 +266,7 @@ size_t encodeLinear(
 	for (i=2; i<dataSize; i++) {
 		ints[0] = ints[1];
 		ints[1] = ints[2];
-		ints[2] = data[i] * fixedPoint + 0.5;
+		ints[2] = static_cast<long long>(data[i] * fixedPoint + 0.5);
 		extrapol = ints[1] + (ints[1] - ints[0]);
 		diff = ints[2] - extrapol;
 		//printf("%lu %lu %lu,   extrapol: %ld    diff: %d \n", ints[0], ints[1], ints[2], extrapol, diff);
